Make side blob lazy loaded
