+++ client/branded/BUILD.bazel
@@ -86,6 +86,7 @@ ts_project(
         "src/search-ui/components/RepoMetadata.tsx",
         "src/search-ui/components/RepoSearchResult.tsx",
         "src/search-ui/components/ResultContainer.tsx",
+        "src/search-ui/components/SearchResultPreviewButton.tsx",
         "src/search-ui/components/SearchResultStar.tsx",
         "src/search-ui/components/SmartSearchPreview.tsx",
         "src/search-ui/components/SymbolSearchResult.tsx",
@@ -158,6 +159,7 @@ ts_project(
         "src/search-ui/results/sidebar/revisions.ts",
         "src/search-ui/results/use-items-to-show.ts",
         "src/search-ui/results/useSearchResultsKeyboardNavigation.ts",
+        "src/search-ui/stores/results-store.ts",
         "src/search-ui/util/index.ts",
         "src/search-ui/util/query.ts",
         "src/search-ui/util/stars.ts",
@@ -199,6 +201,7 @@ ts_project(
         "//:node_modules/ts-key-enum",
         "//:node_modules/use-callback-ref",
         "//:node_modules/use-resize-observer",
+        "//:node_modules/zustand",
         "//client/shared:graphql_operations",
         "//client/shared:shared_lib",
     ],
