+++ client/web/BUILD.bazel
@@ -207,6 +207,7 @@ ts_project(
         "src/codeintel/BrainDot.tsx",
         "src/codeintel/ReferencesPanel.tsx",
         "src/codeintel/ReferencesPanelQueries.ts",
+        "src/codeintel/SideBlob.tsx",
         "src/codeintel/index.ts",
         "src/codeintel/location.ts",
         "src/codeintel/promise.ts",
@@ -1314,6 +1315,7 @@ ts_project(
         "src/repo/blob/codemirror/tooltips/HovercardView.tsx",
         "src/repo/blob/codemirror/tooltips/LoadingTooltip.ts",
         "src/repo/blob/codemirror/tooltips/TemporaryTooltip.tsx",
+        "src/repo/blob/codemirror/types.ts",
         "src/repo/blob/codemirror/utils.ts",
         "src/repo/blob/definitions.ts",
         "src/repo/blob/own/FileOwnershipEntry.tsx",
