+++ client/web/src/repo/blob/codemirror/types.ts
@@ -1,0 +1,4 @@
+export enum SearchPanelViewMode {
+    FullSearch = 'full-search',
+    MatchesOnly = 'matches-only',
+}
